fix: add saldo tarik and topup on admin

+++ app/Http/Controllers/Admin/StudentAchievementController.php
@@ -2,13 +2,14 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Models\School;
+use App\Models\Classroom;
 use Illuminate\Http\Request;
 use Yajra\DataTables\DataTables;
 use App\Models\StudentAchievement;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\Admin\StudentAchievementRequest;
-use App\Models\Classroom;
-use App\Models\School;
 
 class StudentAchievementController extends Controller
 {
@@ -17,15 +18,19 @@ class StudentAchievementController extends Controller
      */
     public function index()
     {
+        if (!Auth::user()->can('Manage Prestasi Santri')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
+
         if (request()->ajax()) {
-            $data = StudentAchievement::with('student', 'academicYear', 'classroom', 'school')->latest()->get();
+            $data = StudentAchievement::with('student', 'academicYear', 'classroom')->latest()->get();
             return DataTables::of($data)
                 ->addColumn('action', function ($data) {
                     $actionEdit = route('student-achievement.edit', $data->id);
                     $actionDelete = route('student-achievement.destroy', $data->id);
                     return "<div class='d-flex justify-content-center'>" .
-                        view('components.action.edit', ['action' => $actionEdit]) .
-                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id]) .
+                        view('components.action.edit', ['action' => $actionEdit, 'name' => 'Prestasi Santri']) .
+                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id, 'name' => 'Prestasi Santri']) .
                         "</div>";
                 })
                 ->rawColumns(['action', 'link'])
@@ -39,7 +44,11 @@ class StudentAchievementController extends Controller
      */
     public function create()
     {
-        $schools = School::latest()->get();
+        if (!Auth::user()->can('Create Prestasi Santri')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
+
+        $schools = School::hasSchool()->orderBy('name')->get();
         return view('admins.student-achievement.create-edit', compact('schools'));
     }
 
@@ -48,6 +57,9 @@ class StudentAchievementController extends Controller
      */
     public function store(StudentAchievementRequest $request)
     {
+        if (!Auth::user()->can('Create Prestasi Santri')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         StudentAchievement::create($request->validated());
         return redirect()->route('student-achievement.index')->with('success', 'Prestasi siswa berhasil ditambahkan');
     }
@@ -65,7 +77,11 @@ class StudentAchievementController extends Controller
      */
     public function edit(StudentAchievement $studentAchievement)
     {
-        return view('admins.student-achievement.create-edit', compact('studentAchievement'));
+        if (!Auth::user()->can('Edit Prestasi Santri')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
+        $schools = School::hasSchool()->orderBy('name')->get();
+        return view('admins.student-achievement.create-edit', compact('studentAchievement', 'schools'));
     }
 
     /**
@@ -73,6 +89,9 @@ class StudentAchievementController extends Controller
      */
     public function update(StudentAchievementRequest $request, StudentAchievement $studentAchievement)
     {
+        if (!Auth::user()->can('Edit Prestasi Santri')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $studentAchievement->update($request->validated());
         return redirect()->route('student-achievement.index')->with('success', 'Prestasi siswa berhasil diperbarui');
     }
@@ -82,6 +101,9 @@ class StudentAchievementController extends Controller
      */
     public function destroy(StudentAchievement $studentAchievement)
     {
+        if (!Auth::user()->can('Delete Prestasi Santri')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $studentAchievement->delete();
         return redirect()->route('student-achievement.index')->with('success', 'Prestasi siswa berhasil dihapus');
     }
